Allow late-joining websockets to get program output.
Allow late-joining websockets to get a program's output even after a
job ends by adding a configurable delay before actually removing a job
from the job manager's set of jobs. This can be changed by
juttle-config's juttled.lingering_job_timeout option, and defaults to
10 seconds.

During the lingering timeout, the job is marked as stopped. When
adding an endpoint to a job, do a second check to see if the job is
stopped. If it is, replay all messages, send the job_stop message, and
close the connection.

This revealed a possible race condition where a endpoint.close() could
be called before all messages have actually been sent. To address
this, add a flag endpoint:_closing which is set to true in
endpoint.close() if there are queued messages. In drain() once the
queue is empty it calls endpoint.close() again.

Add a unit test that has a websocket join after the program completes,
but before the lingering timeout kicks in.