Address feedback from PR.
Make immediate running orthogonal to running from a bundle or
path. You can choose to run either a path or a bundle and either wait
for the results or not. Update docs to reflect this.

This streamlines the two paths in job-handlers a bit. Now there's a
single function create_job that either bundles or not, and then calls
run_program with wait and timeout set from the request.

Change wait_for_end to wait.

Also change how the promise chain is created to push it into the
XXXJob class. Everything is now handled by job.start(). For the simple
case implemented in JuttleJob, start() just starts the program and
returns a promise with the job id and pid. ImmediateJuttleJob builds
on this to additionally wait for the program to finish and then
resolves with the output of the program. (Remove results() while doing
this, it's not needed any longer).

Add minimal coverage of the additional combinations (bundle + wait,
path + no wait). Also set wait: true for all of the old path combinations.