Add client program.
Add client program bin/juttle-engine-client, which inherits all the
capabilities of juttle-service-client and additionally has the
following commands:
 - browser --path <path>
 - push --path <path> [--topic <topic>]
 - watch --path <path> [--topic <topic>]

Thanks to the changes in juttle/juttle-service#22, this is easy as
defining some commands/cmdargs and implementing the commands in
functions. The functions were all copied from the latest
outrigger-client.

Also add the endpoint notifier back and add a handler for websocket
connections on /rendezvous/:topic to pass websocket connections to the
endpoint notifier. I also switched endpoint notifier to ES6 classes.

Add all necessary new packages to package.json. Although when I
tested, I used the master github version of juttle-servive, I didn't
check in that change. We'll need to update package.json to an
appropriate version of juttle-service.