Update outrigger to juttle-engine.
Update outrigger to juttle-engine in all READMEs, Dockerfiles, and
docker-compose files.

Simplified the README for the Dockerfile to not
talk about how to run the container. That's now better covered in the
examples README. Just point to that README instead.

Remove the example programs below docker/images/outrigger (now
juttle-engine), the examples in examples are better.