Add AWS and Cloudwatch example programs.
Add a collection of example programs for the AWS and Cloudwatch
adapters, copied over from the jut 1.0 built-in boards for the aws
integration.

They are grouped into the following categories:

- overview (overview.juttle): shows the number of EC2 instances, RDS
  total allocated storage, EBS total Iops, as well as a list of
  "problem" items for EC2, ELB, and Lambda.

- demographic (demographic.juttle): show an aggregate as a timechart
  and a demographic as a piechart and timechart for a given product,
  with overlaid events any time the set of items changes.

- capacity_planning/*.juttle: this is broken into multiple programs
  for various products, and provides different ways to combine
  cloudwatch and aws information to give an idea of the reserved
  resources vs resources actually in use.

To make it easier to view aggregate/demographic information across
juttle programs, create a mysql database "aws" with table
"aws_aggregation" and populate it from a juttle program that
periodically fetches aws information, aggregates it using the
aws_module.juttle file (currently pulled directly from github, will
replace with an adapter provided module once juttle/juttle#325 is
fixed), and writes to the mysql database.

dc-aws-cloudwatch.yml contains the mysql container, an additional
mysql container that runs the script to create the database/table, and
an additional juttle-engine container to run the poller juttle
program. The startup of these tasks is a bit hokey with some sleeps to
make sure everything is started in the right order. docker-compose 1.6
claims to support the notion of 'events' where containers can signal
their readiness to each other, but 1.6 was only released days ago so
it's a bit bleeding edge to rely on imho.

Without this, all the example programs take ~10-20 seconds to get any
aws aggregate/demographic information which has to be fetched from
scratch every time a program is run. With the mysql database + poller
script, as long as the docker-compose environment is running you get a
steady stream of updates that can be used across programs.

I'd like to add some cloudwatch-specific programs that allow you to
view cloudwatch-specific metrics, but that's really awkward to use
without inputs with -juttle, so I'll skip that for now.