Add a dockerfile that builds from source.
Add a Dockerfile that builds from a checked out tree, loosely
following the steps implied by
https://github.com/nodejs/docker-node/blob/master/4.3/onbuild/Dockerfile:

 - copy package.json to image
 - run npm install within image
 - copy source code to image

Continue to use /opt/juttle-engine/juttles as the root for consistency with
the released image.

This npm install does not install globally, so symlink
juttle/juttle-engine to /usr/local/bin so they can be run without a
path.