Add topic msg validation
On the client end of things we need to ensure that every topic msg
includes a bundle_id along with the bundle_object. The bundle_id is used
to tell whether the incoming message is an update to a bundle or a new
bundle altogehter. All this effects the way we run the bundle and its
inputs.

Added validation to topic msgs to ensure that `bundle`, `bundle_id`, and
`type` are included in any client requests. If not, an error message is
returned to the sending client.