Add example programs for reading nginx logs.
Add example programs and docker-compose files that show how to read
nginx logs using logstash, save the parsed results in elasticsearch,
and summarize the parsed results in juttle programs.

I am checking the programs in, but I'm not including them in the
examples README or providing a standalong README for now, as the
program seems pretty specific to our situation. Maybe we need another
repository just for internal-facing programs and supporting
infrastructure?