Add cloudwatch-specific example programs.
I was going to wait for juttle-powered input support to be in
juttle-engine before adding cloudwatch-specific programs, but the set
of metrics isn't *too* horrible and I've enumerated them anyway in the
documentation, so add the programs now.

aws_defined_metrics.juttle names the cloudwatch metrics that exist for
each product. That in turn is used for the inputs aws_control_metric,1,2.juttle.

The programs are:
- top_items.juttle: Show a barchart of the top 10 items by any
  cloudwatch metric.
- metric_timechart.juttle: Show a timechart of a given cloudwatch
  metric across all items.
- metric_outlier_timechart.juttle: Like metric_timechart, but shows
  the difference from the median as a way to highlight outliers.
- metric_correlation_scatterchart.juttle: Show a scatterchart of any
  cloudwatch metric vs any other cloudwatch metric for all items.

Also add a barchart of the top EC2 instances by CPU Utilization to the
overview page, just as a way to show off cloudwatch support.