Improve outlier behavior with 1 minute sampling.
I was seeing bad results with 1 minute batching for some metrics, so
increase the window width to 30 mintues and the period to 5 minutes.

Also go back to the original behavior wrt setting oldtime before the
join--this was actually necessary. I think I can fix this by using the
appropriate option to join, but I'll fix that in a separate PR.