Fix browser command with non-default root path.
Make the browser command easier to use when run against a
juttle-engine/juttle-service with a root path other than /. Do this by
fetching the server's root directory and comparing that to the path
provided in the --path argument. If the root directory isn't a prefix
of the --path argument, throw an error. Otherwise, take just the part
not covered by the --root argument and use that as the ?path= cgi
param in the url.

In juttle-engine-client, also fully resolve the path immediately
instead of several times in the function for each command.

Also update juttle-engine to rely on juttle-service's built-in
defaults with juttle-config.js and command-line overrides. This
simplifies the initialization steps a bit.

These changes caused the test coverage numbers to dip slightly below
the thresholds, so provide some additional options when creating the
engine to cover one additional branch, which got the coverage numbers
up again.