examples: tweak npm download counts to not emit warnings
Some odd interactions with mismatched intervals started happening
when the new week began at 5pm today that caused nulls to be generated
from the moving window reducer that triggered warnings in the program
output.

Update the program so that we pass the interval along properly to the
trend sub and the views so there are no more warnings.